Support resuming knitted projects and bugfix
Added resume_knit that triggers if called as 'sew resume needles', or with knitting needles in hand.  Analyzing a pair of knitting needles with something on it gives like 40 lines of scroll and TWO separate RTs from one action, which screws up bput entirely.

Tested on knitted items resumed at each stage (push/knit/turn/cast) and on empty needles.

Also, encountered a bug of some sort with the PUSH command in caravans.  Pushing in a caravan while someone is online with a name beginning with 'My' always results in 'Myfoo is not here!'

If it runs into that case it subs out 'my ' and just becomes 'push needles'