Toggle hide and add fare withdrawal to bescort (#3411)
* Toggle hide and fare withdrawal to bescort

Two changes, and I'm going to set this out there for a while to do testing and get input, since bescort is so important.

First, spoofed hide? in bescort which does DRC.hide? unless bescort_hide is false.  bescort_hide going into base.yaml as true by default for now.

Second, added get_fare?.  If bescort_fare_handling is true (default false for now) it will check to see if you're wanted, and if not it walks to the nearest bank to withdraw fare, then returns to the room it started in.

I've tested this up and down on the Crossing/Leth ferry and it works fine.  I'll want to test it everywhere.

Also, changed default 10_000 on dirigible fees to circle*50.

Closes https://github.com/rpherbig/dr-scripts/issues/3408

* Change method for getting money

Swapped to get_money_from_bank and using minimize coins to handle fare more exactly.

Check justice after arriving at the bank in question and does not withdraw if it detects you're wanted for being a bad man.

Fixed a Mer'Kresh apostrophe error.