Generalize anvil-cleaning and fix logbook case
Someone or someones have been putting things other than bronze ingots on anvils down in Shard to screw with crafters.  This is meant to generalize the clean_anvil logic.

Examples of how it responds to found items below:
```
[smith]>look on anvil
On the iron anvil you see a tiny coal nugget.
>
[smith]>clean anvil
You glance at the anvil's contents, and realize you should be able to remove them yourself.
>
[smith]>get nugget from anvil
You get a tiny coal nugget from atop an iron anvil.
>
[smith]>put nugget in bucket
You drop a tiny coal nugget in a large waste bucket.
>

[smith]>look on anvil
On the iron anvil you see a bronze ingot.
>
[smith]>clean anvil
You glance at the anvil's contents, and realize you should be able to remove them yourself.
>
[smith]>get ingot from anvil
You get a bronze ingot from atop an iron anvil.
>
[smith]>put ingot in bucket
You drop a bronze ingot in a large waste bucket.
```

And tested with an actual clean anvil as a control.
```
[smith]>look on anvil
The anvil's surface looks clean and ready for forging.
>
--- Lich: forge active.
--- Lich: buff active.
```

Also there's a problem I ran into frequently while crafting containers.  Trying to PUT LOGBOOK IN RUCKSACK when you have a rucksack bundled with your logbook makes the parser get really confused.  In the case that it can't find that container it just stows the logbook instead.

I've been running with that change for about 4 months of crafting, so I'm pretty confident in it.