register.lic - updated get logic
Changed the logic for getting the register to try again if it isn't in your crafting container.  Now it tries your crafting container first, and then tries to just get your deed register.  If that fails, it messages as such exits.  Otherwise it continues to search the register and then puts it in the right container.