common methods for instruments (#3546)
Added play_song? to common.lic to reduce dependency on ;performance which has caused a number of issues.

play_song? defaults to a worn_instrument, but tested well with held instruments if given worn false.

Little was changed from ;performance's play_song, which was used as a base, except...

worn_instrument: added to base.  Defaults to zills, but can be made to be be cowbell, nose-flute, whatever.

Some unecessary returns were removed from play_song.  play_song? itself returns true if it succeeded in starting to play a tune of the right type, and false if it didn't get what it needed.

clean_instrument also works for worn/nonworn instruments, but other than letting instrument = worn_instrument or instrument little was changed.  It won't exit if it can't find a shamwow now, since anything could be calling it.  Instead, if it can't find a cloth then clean_instrument will return false and keep playing.  Any script that wants to use it can account for that.

I've been running this with performance calling these methods on the roads for a while, and haven't had any problems.  So far it worked with zills and cowbells, and a non-worn instrument, and it's cleaned all three.