almanac.lic: Add support for the droghtmans almanacs
These can be turned to individual skills. This adds support for a yaml setting to pick which skills to use, and it will pick whichever skill is at the lowest mindstate to study. If none of the skills are under 18/34 it will loop every 20 seconds checking to use it when one is.

Skills to use can be customized like

almanac_skills:
- Mechanical Lore
- Outdoorsmanship
- Scouting

Fixes #3559