Update to afk.lic for grabbing at foot ammo
The get step is not needed and due to a bug with the new Droughtman's blunt-tip arrows, afk will not grab them if you have. Individual arrows show up as "arrows" instead of "arrow" and if you "get my arrows" it will grab the arrows on your person and get stuck in a loop.