mining-buddy - updated get and store
It was always searching to see if there is a shovel on the forging belt even if the person specified their mining implement to be a pickaxe.  This caused it to try to untie the pickaxe from the forging belt if the person had a shovel on their belt.

Now it searches for whatever your mining implement is.