common.lic - add set_stance method
Method to set stance.  Adding it to common because it makes more sense here than as a helper method in common-arcana.

When used for parry or shield stance, it sets evasion to 100 and the other to 0.  This is for instances when either shield or parry can't be used for some reason (took off shield for lockpicking, bow in main hand, etc).

It uses all points for evasion stance.