Common and common arcana update (#3698)
* common.lic - add set_stance method

Method to set stance.  Adding it to common because it makes more sense here than as a helper method in common-arcana.

When used for parry or shield stance, it sets evasion to 100 and the other to 0.  This is for instances when either shield or parry can't be used for some reason (took off shield for lockpicking, bow in main hand, etc).

It uses all points for evasion stance.

* common-arcana - updated to use new set_stance method

Now uses set_stance method from common.lic

* test_common-arcana - update to work with new version of common arcana

And the new helper method in common.lic

This should fix the Travis CI failures.  Thanks to @rcuhljr for the code!