rebase attempt 2 (#7)
* rezz - correction to infuse_amount method (#3699)

It was returning rezz['mana'] instead of the calculated values by mistake.

* Common and common arcana update (#3698)

* common.lic - add set_stance method

Method to set stance.  Adding it to common because it makes more sense here than as a helper method in common-arcana.

When used for parry or shield stance, it sets evasion to 100 and the other to 0.  This is for instances when either shield or parry can't be used for some reason (took off shield for lockpicking, bow in main hand, etc).

It uses all points for evasion stance.

* common-arcana - updated to use new set_stance method

Now uses set_stance method from common.lic

* test_common-arcana - update to work with new version of common arcana

And the new helper method in common.lic

This should fix the Travis CI failures.  Thanks to @rcuhljr for the code!

* Rubocop

* Update thieving data

* appraisal.lic - allow appraisal of closed pouches (#3703)

Adds the ability to open pouches that are closed but not tied so that they can be appraised.

https://github.com/rpherbig/dr-scripts/issues/3696

* Add special glove remove message (#3705)

Some crazy gloves from the festival get taken off with the message:

"Without any effort, the mail gloves slide themselves off of your hands.  The silvery sigils on the backs of the gloves flicker once, then fade slowly."

* Add rare gems to sorting yaml (#3706)

* Fix for crashing in column mode (#3707)

* Fix for sorter crashing in columned mode (#3708)

fix for the fix