Create tithe.lic (#3724)
This script breaks out the tithe functions from theurgy and crossing-training, allowing them to be run as a standalone script (for example, for use in t2). It additionally includes praying to Chadatru if a paladin to bolster soul state. As the two actions are on a similar timer, coupling them makes sense (but oth actions are tracked separately, just in case).