bescort.lic - Update to deal with fog in the river (#3738)
This should handle all possible fog related issues.  It deals with the following:
There's fog in any of the middle rooms of the river.
You got pushed east into one of those rooms that goes to that slope in crossing.
There's fog in either of the endpoints and you've just reached it.