Move more stuff into common play (#3740)
I left in the middle of this process, but it's working out for me so far.

Adding a few more universal items from performance, as well as skip_clean.

Right now, my ;performance has an entry to remove hand items, calls play_song?, then goes into the watch loop until it's done.  I bet that behavior could be put into other scripts with a flag instead, and I'll try it in ;trade.

After this goes live I can at least add the performance changes so it'll support zills and cowbells.