[combat-trainer] Allow wearing bundles on other locations than shoulders
The default location for wearing a lumpy bundle is the shoulder, but
it's a fairly in-demand slot. This change adds a new skinning option to
`PULL MY BUNDLE` a number of times, allowing it to be worn on the
shoulders, back or waist as well.

Example:

```yaml
skinning:
  skin: true
  pull_count: 1
```

This will pull the bundle once, draping across the shoulders.