athletics - remove duplicate logic - add cheap-rope support (#3951)
* athletics - remove duplicate logic - add temp-rope support

This shouldn't be merged yet -- the common change was fairly recent.  Also, I want to find someone with a permanent climbing rope to test this, if possible.  I'm not happy about the fput in check_rope, but I don't actually know what a real rope would look like, just how a limited one looks when it fails.

First, moved stop_play, clean_instrument, and play_song? out to common.

Second, added a flag to kill the training and quit out if your climbing rope runs out of minutes.

Third, added an arg to just stay in your room and not move with a climbing rope.  Mainly this is so I can climb my rope inside or behind my caravan.

* Adjust behavior and take a guess at climbing diff

* be less generous with guessing

* change fput