Create tarantula.lic
Tarantula script!  Yaml settings below:

This one is only necessary if you want to set it to true.  The script defaults to false.
tarantula_debug: false

This one is only necessary if you got yours altered to be something other than a tarantula, or you want it to use something other than the full name "biomechanical tarantula" when using it.
tarantula_noun: tarantula

This one is to tell the script which skills to use per skillset, in and out of combat.  You can specify the same skill both in combat and non_combat.
tarantula:
  Armor:
    combat:
    - Chain Armor
    - Defending
    - Shield Usage
  Weapon:
    combat:
    - Melee Mastery
    - Missile Mastery
  Magic:
    combat:
    - Utility
    non_combat:
    - Sorcery
  Survival:
    combat:
    - Evasion
    - Skinning
    non_combat:
    - Athletics
  Lore:
    non_combat:
    - Appraisal
    - Performance
    - Scholarship

This is for skills that you NEVER want the tarantula to use.
The script will try other skills if nothing is available from the ones you specify in the above setting.
tarantula_excluded_skills:
  - Forging
  - Outfitting
  - Alchemy
  - Enchanting
  - Engineering