Added Crossing Estate Holder Gargoyles (#4108)
Should not be approved until the map upload with a stringproc for room 1629 is live.  Room data does not transition from 1629 to 10732 but does transition from 10732 to 1629.  Room 1629 appears to be the problem as the entire room is nothing but procs.
Once approved, this is a super yummy new empath-safe hunting spot we are very excited to be use!!!