Added command line argument overrides for entry, room and loot_type (what to do with loot)
Added support for multiple rooms

Migrated reused settings (stealing_bag, use_lockpick_ring, lockpick_container) into burgle settings due to some confusion as to the location of those settings.  If users want to reuse these settings, yaml anchors can provide that functionality, and if not they can set different settings. - This is a non-breaking change as there is code to mitigate that and warn users of the change.

Changed how dropping loot works.  It will always attempt to store it now, meaing you must always have a burgle_settings:loot_container set to work.

Added flags to watch for footsteps, and if safe_mode setting is set, try to get out as  quick as possible.

Add cycle option to pick from lowest of locksmithing/athletics learning rate, and lowest skilled if learning rate are the same.

Add checks to getting to your burgle room so if it fails, it exits out.

Added before:/after: so that custom scripts for things such as putting pets away/taking them out can be done