Add "fill pouch with box" (#4337)
* Add "fill pouch with box"

Added "fill pouch with box" line immediately after box opens and before it LOOKs to read the contents.  Reduces the number of actions needed and dramatically lowers the chance of dropping a gem that didn't parse correctly.

* update wear pouch to include the adjectiv

How this has never been addressed, I don't know but adjective should always be used.  I managed to drop a full pouch today because of trying to mess with the wrong pouch.