Added fix for simu "missing" a game command.
After much testing and discussion, the error seems to be in the XML.  Several people have been using this fix for months with no hangups and a near-elimination of the issue.  In cases where the weapon does load correctly the first time, the "already loaded" match passes smoothly to the next action.