[script] [dusk-labyrinth] efficiently walk rooms (#4408)
* [script] [dusk-labyrinth] efficiently walk rooms

Builds an internal map of the maze to efficiently skip rooms that the player has already been to (e.g. back tracking to pick a new direction). This saves several minutes by not spending time waiting for a pet scurrying message when we know the pet won't be found in that room.

When maze reshuffles then script clears its knowledge of the maze and rebuilds it anew.

* add comments