[script] [common] New method: can_see_sky? (#4450)
* can character see the sky?

Discord Lich thread: https://discordapp.com/channels/745675889622384681/745676101392793651/746953119006588980

Answers the simple question: can I see the sky?

Some abilities, such as casting Lunar Magic spells or doing Moon Mage celestial observations, require seeing the sky. This new function is a easy way to check if you can see the sky or not then take appropriate action. For example, to not whip out a telescope and try to observe the sky that you cannot see.

* fix to avoid false positives

Ironically, testing this in room 5713 (Empath's guild in Crossing that is inside with view of the sky) gave a false positive when the messaging the game first saw was someone getting an item from **inside** a container. Matching on the one word `inside` was too permissive.

```
> ,e echo DRC.can_see_sky?

--- Lich: exec1 active.

[exec1]>weather

Hawqwinter gets some holy water from inside his backpack.
> 
[exec1: false]

--- Lich: exec1 has exited.
```