optimize checking for sky
If you stay in a room for an extended  time and you can't see the sky then the game window is spammed every 10 seconds with the output of the `weather` command. This commit ensures you only check the weather once while in that room if you couldn't see the sky. If you move to another room then the weather checks will resume.