changing the waitfor to pause
Getting reports of mismatching this morning - the waitfor assumes  you just started playing when there's a few reasons why that may not be the case.  I think a short pause that allows you to get your hands/instrument sorted is less of a stumbling block.  I would consider it an edge case to not be using a worn_instrument these days.