More options to check which hand holds an item
* Is item in the left hand?
* Is item in the right hand?
* Is item in either hand?
* Does each hand hold one of the items?

One use case is when deciding if to swap a weapon between hands to train offhand weapon skill. If the weapon is already in your left hand, you don't need to swap.

Or, perhaps an action requires the item in a particular hand, you can check if it's already there or not with these new utility functions.