Match on full words from config
Without the leading `\b` in the regex, then words that end with the adjective match which may not be accurate. For example, "oak spear" is in your config but "redoak spear" would also match although it's a wrong match.

Without the trailing `\b` in the regex, then words that begin with the name match which may not be accurate. For example, "leather book" is in your config but "leather booklet" would also match although it's a wrong match.