bescort - added flying broom/carpet use to faldesu and segoltha
I added functionality to use flying mounts and set it up to work for both the segoltha and faldesu rivers.

You need to add flying_mount setting to your yaml.  Here's an example with one of the new flying carpets.
`flying_mount: silk carpet`

I updated the map today and it should be pushed through to everyone by tomorrow or day after.