bescort - added flying broom/carpet use to faldesu and segoltha (#4512)
* bescort - added flying broom/carpet use to faldesu and segoltha

I added functionality to use flying mounts and set it up to work for both the segoltha and faldesu rivers.

You need to add flying_mount setting to your yaml.  Here's an example with one of the new flying carpets.
`flying_mount: silk carpet`

I updated the map today and it should be pushed through to everyone by tomorrow or day after.

* Update bescort.lic

Updated to skip all athletics checks for the segoltha

* Update bescort.lic

* base.yaml - Add flying_mount setting

Added the blank `flying_mount: ` setting to base.  It's currently only used in bescort.