[script] [equipmanager] Fix success/failure match strings for transforms (#4637)
* fix success/failure match strings for transforms

In get_item_helper, it checks for exhausted then failures then successes match strings, in that order. Since the name of the transformed weapon was in the failures list then on successful transformation the script thought it failed and went through the failure recovery logic.

* handle auto-worn transformed weapons

For example, when "shake exoskeletal axe" then it's auto-worn as an armband. Therefore, stow_helper needs to handle when the item it tries to wear is already worn.

* wait for any RT on transforming or retrieving item