play nice with others
### Background
* The `study-art` script runs `performance` in the background.
* You may have other scripts running, too, like `sanowret-crystal`.
* This script when looping over items to study in the room can dominate your roundtime, blocking other background scripts from having a turn.
* One impact is that once you finish a song, you may not be able to start another song.

```
> 
[study-art]>study painting

[sanowret-crystal]>stow my sanowret crystal

[performance]>play battle on my zills

Framed in a magnificent rectangle of rich mahogany, the bright bedroom reveals an Elven couple standing to face the viewer.  The woman has red-gold streaked blue-black hair, and she looks downward, seemingly demure.  Her bright red silk dress pools at her feet like newly spilt blood.  The darkness of the material casts a stark contrast to her very pale skin and serves to make her look frightened and childlike.  To her left, the man is painted with piercing blue eyes and dark brown hair, and he looks out  regally as if he has just been crowned prince.  Either by design or accident, the dark outfit gives the man a sinister air that is weakly counteracted by the red of her dress.  Behind the couple, in the very center of the painting, is a round seven-pointed mirror.  On closer inspection of the mirror, the painting shows such intricate detailing that one can see the backs of the couple.  Beyond them is the artist, who shares a face with the bride, leaning forward to paint on her canvas.  A small golden plaque underneath the painting reads, "'The Ja'Haadraan-Daemondred Marriage' by Ysselt Ja'Haadraan-Daemondred."

You deduce the wood is in fine condition and displays expert craftsmanship and exquisite detail.
You deduce the paint is fully saturated and is protected from further wear for the time being.
You deduce the canvas is obviously taken care of and displays the signs of an elementary attempt at restoration.
You deduce the piece probably weighs around 150 stones.
You deduce the style of art is similar to that of other works crafted during the Modern Times in and around the province of Zoluren.
Roundtime: 30 sec.
> 
...wait 30 seconds.
> 
...wait 30 seconds.
```

### Changes
* Introduce a 2 second pause between studying items to give other scripts a chance to do something (e.g. restart song, gaze crystal)

## Tests

### with pause, can restart a song
```
You deduce the style of art is similar to that of other works crafted during the Time of Chaos in and around the province of Therengia.
Roundtime: 30 sec.
> 
You continue playing on your thin-edged zills.
> 
Portoi just arrived.

> 
You finish playing a quiet battle song on your thin-edged zills.
> 
[performance]>play battle on my zills

You begin a quiet battle song on your thin-edged zills with only the slightest hint of difficulty.
> 
[study-art]>study carving

A single crystal rose sits....
```