Update Pause room for Undergondola route
How we've never come across this, I don't know. I suppose it must be because critter spawn in the area is sparse at best. The room where we wait between rounds and do outdoors on the "undergondola" route was in a la'tami room.  Being in combat results in not being able to collect which in turn causes the script to begin climbing before the internal cooldown on climbing gains.  This means you waste time climbing when you can't learn and are also not collecting for outdoors.  New room 9515 is a safe area I've used for years and know it to be free of mobs only four rooms away from the original.