athletics.lic - added outdoorsmanship method
Added a method to call outdoorsmanship between climbing runs.
It consolidates the code and also fixes the issue where your outdoorsmanship is already over 29/34 which stops the outdoorsmanship script, causing you to start the climbing loop again immediately.  This will now pause the remainder of the time so that you don't make a climbing loop and learn nothing.