register.lic - Updates for labeling blank pages and reading someone else's register
Updated the script to message appropriately and put the register and stylus away if you try to label a blank page.

Updated the search method to deal with reading someone else's register.  It no longer worked even though the messaging hadn't changed so I moved the checks from the loop to the bput for reading.