[scripts] [common-items] [02] Add section header comments
This is the 2nd of many PRs I'll be submitting for common-items. It depends on https://github.com/rpherbig/dr-scripts/pull/4930

To make it easier to review the changes, I'm phasing them in.

This one focuses on comment sections in the file to help separate logical methods (e.g. methods that GET things vs. OPEN things vs. COUNT things, etc). Two methods were moved to better align with their section.
- `exists?` moved to "Search for Item" section
-  `get_inventory_by_type` moved to "Check Container Contents" section