[scripts] [common-items] [04] get_item?
This is the 4th of many PRs I'll be submitting for common-items. It depends on https://github.com/rpherbig/dr-scripts/pull/4932

To make it easier to review the changes, I'm phasing them in. However, because I can't open a PR in this repo against my forked branch of https://github.com/rpherbig/dr-scripts/pull/4932, until that dependency is merged then this PR's diff is noisier than it really is. Instead, please review the net new commit 

This one focuses on moving the match strings for the "get item" methods to $DRCI constants for easier maintenance and readability as is done with other methods in this class.