[script] [common-items] Refactor $DRCI constants to class variables
This is the 8th of many PRs I'll be submitting for common-items. It depends on https://github.com/rpherbig/dr-scripts/pull/4936

To make it easier to review the changes, I'm phasing them in. However, because I can't open a PR in this repo against my forked branch of https://github.com/rpherbig/dr-scripts/pull/4936, until that dependency is merged then this PR's diff is noisier than it really is. Instead, please review the net new commit 

This one focuses on refactoring $DRCI constants to class variables per this feedback https://github.com/rpherbig/dr-scripts/pull/4930#issuecomment-830129449