Temporary Map Fix (#5116)
Today Simu updated repair shops and broke their titles so ;go2 cannot function.  The map is currently locked and  has been for months with no solution in place.  This is obviously not an ideal fix but it will at least get folks moving when they are otherwise at a standstill.  If there is a place that would be better, I'm happy to edit.  @rcuhljr  @rpherbig   

Otherwise if this is good, can we get it pushed ASAP?