Add New Dissect
Separate from Necro ritual dissect, more closely tied to skinning and available to everyone.  Has been tested by quite a few folks and this is the best usage I can achieve with the Simu-side syntax being super wonky.  Koror said they'll look at the targeting so we may have updates in coming days but people are very anxious to get started and this is functional for current conditions.