[data] [base-hunting] Separating Leth dryads
The twisted dryads (36) and dryad priestesses (34) spawn in different areas and have different rank ranges. It could be useful for people dealing with the low-level hunting ladder to start at 34 and move up to 36 later. This change leaves the 'dryads' zone in place, and moves 4 of those rooms into a separate 'dryad_priestess' zone.