Revision of leth dryads change
This update to the pull request restores the 4 dryad_priestess rooms to the original 'dryads' zone, so that they will still be available as part of 'dryads' but can also be specified by the new dryad_priestess zone. The purpose of this change is to accommodate characters who might end up using those rooms to due minimum mana threshold settings.