[script] [combat-trainer] berserk training_abilities
This adds berserk Landslide and berserk Avalanche as training abilities for warding and utility.  I find this useful on a smaller barbarian who does not gain much field exp from the research options.  It's preferable to have this feature rather than rely on buff_nonspells: which does not respect inner fire thresholds or field exp targets or roundtime conflicts.  Avalanche and Landslide are accessible early trainers which may become staples when the upcoming barbarian mastery split is live (though maybe we'll all be using Tenacity for warding early on).

If this makes sense, seems useful, and isn't obviously broken in some way, I'll look for people willing to test.