NEW SCRIPT - sort-scrolls.lic
This is an alternative to stack-scrolls because that script requires that you have different nouns for each case.

I've written this script so that you have all of your cases with the same noun, but you have to put a wax label on each of them and label them with guild names.  Currently the search is case sensitive, and spaces work so they need to be labeled like this: 
"Bard"
"Warrior Mage"
etc.

There is also a 10th case needed and that one should be labeled "Extras" for overflow if any of the other cases get full.

Here are all of the yaml settings used in the script:

```scroll_sorter:
  stacker: scroll folio
  stacker_container: haversack
  scroll_nouns: ['cerulean scroll', 'glittery-pink scroll', 'silver-stippled scroll', 'gold-stippled scroll']
  close_container: false
```

`stacker:` is the noun or adjective and noun of the case type you're using.
`stacker_container:` is where you have your cases stored, and the script will put them back there as well.  If you don't specify this setting, it will use `default_container:` instead.
`scroll_nouns:` is where you specify any custom scroll types that come from different cases and also from custom papers you can use.
`close_container:` will close the container when the script ends if you set this to true.  If you don't include this setting, it won't close the container.