Add arg to check song and exit
Many people use UserVars.song to train performance in a variety of scripts these days but still have to run ;performance from time to time in order to check their song choice.  This allows people to use the common method (which is awesome) and exit immediately instead of having to sit through an entire song first.