Update to allow Forage Toggle
With new foraging items released recently, it much more effective to train with `forage` instead of collect but it is largely dependent on a trash bin nearby.  It is also quite handy to be able to change your room/item/method on the fly based on where you might be so it can be set by yaml or by args.  All args work independently or together in any combination so long as they're sent in correct order.
Defaults mirror existing behavior
```yaml
# Define as `forage` to override outdoorsmanship gathering method.  Leave blank to collect.
outdoors_method: 
# If you choose forage method, you'll want a room with a trash bin!  Leave blank to not set a default room.
outdoors_room: 
# If you choose forage method, you'll want to select a custom item for best learning rate.
forage_item: rock
```