Refine chart selection equation
There was a hole in the last logic at precisely 1300 but I've had a long time to reevaluate the chart selection method and this version is much less complicated.  Anyone with current modifiers will be uneffected and those relying on dynamic selection should only notice small changes to what charts are chosen.