adding magical research scripts
Thanks to Crannach the original creator of researcher.lic on the in-game repo
Modified to add in all missing research options for normal and symbiotic research. Also added in guild specific research.

check-symbiosis.lic relies on research.lic and will make sure your symbiotic research is both active and the one you want.