Update Shard athletics options
Updating the notes for clear reading with Genie or other front end.  Also re-adding the logic for correctly selecting wyvern mountain when running `;athletics` without an arg.  The way it was previously forced during `undergondola` was the only way to automatically select the correct option.  Relocating that logic to shard_athletics reverts default behavior while still allowing users to block the DC mountain.