Re-added 'random' functionality
When merging the two scripts I accidentally dropped the 'random' function from the sigil argument.

Entering 'random' instead of a sigil name will now randomly select a sigil and then run the specified number rooms based on the roomcap argument.

Further, 'random' will also be updated each loop if the 'loop' option is chosen for roomcap. Otherwise, the randomly chosen sigil will be static for the remainder of the script's runtime.