Rework of stall fix
Due to the order in which certain messages appear I switched the stall fix over to a new flag that listens for the messaging indicating there are no more sigil improvements. This flag is checked each time sigil_info is run, and exits the function after clearing the improvement arrays. An empty improvement array indicates to other functions to run another 'perc sigil improve' which will generate a new list of improvements.