Script Modernization
I've added the suggestions from the group (all awesome). I've so far run into no issues with this script. Please review but it's very close to finished unless there are any major issues.

I'll be testing for the next 24 hours or so, might need to make a few more tweaks.