Update workorders.lic
Assuming this was added at some point to integrate using techs to repair your own tools? If anything needs to be repaired, this adds 20-25 seconds to the job. If nothing needs to be repair, this saves 20-25 seconds, but in my experience, even with only two items crafted, the former occurs far more often than the latter. Moreover, repairing every time saves on repair times down the line.

Furthermore, the check of whether to repair at all in the yaml occurs before this method, so nothing is lost function-wise.