[data] [base.yaml] Add handling for Darkling Wood tree
The XML stream for this tree nests the `<nav/>` tag into the body of the room description, so movement isn't detected correctly. This slows things down a fraction of a second, and adds a double look (a single look wasn't enough...) so you don't go too fast, and fall off. This takes my once-nightly fall, to not falls at all.