sort-scrolls - fixes and updates
The code for finding a specified spell to remove was broken:
`;sort-scrolls find bless`

```--- Lich: error: undefined local variable or method `spell' for #<ScrollSorter:0x04518ae0>
Did you mean?  spells
	sort-scrolls:148:in `find_spell'
	sort-scrolls:35:in `initialize'
--- Lich: sort-scrolls has exited.```

This is due to a few reasons.  The script is searching for  `@all_spells[spell]` instead of `@all_spells[name]`. 
Also, the argument that gets read in is stored as all lower case and the data from base-spells is capitalized appropriately.

To fix this I re-added the `find_guild` method and updated it to use `casecmp()` to compare so capitalization doesn't matter.
This also allows this method to be used by both the `find_spell` method and also the `check_scroll` method.
I changed the way the data is stored back to an array rather than storing the whole hash from base-spells.  This allows me to ignore cantrips and fix AP spells as I pull the data in.

I added a new message if you try to find an AP scroll in your collection.

I updated the `find_spell` method to search the Extras case if it doesn't find the spell in the appropriate Guild case first.

Updated the descriptions for the arguments to clarify how to search for spells that are more than one word.