[data][base-recipes] advanced AP ritual focus fix
Odd one, maybe a one off.

This recipe requires a secondary sigil to begin, then ANY sigil to complete it.

The issue is that as is, the script purchases congruence, but then looks for decay (secondary) to begin.

By adding enchant_stock1 to Decay, and enchant_stock2 to Congruence, this work order can be completed as expected.

Technically this PR could of been completed by just adding sigil 10, but I wanted to change the order to reflect SIGIL and #1 and Congruence (or any) as #2.