tarantula - Fixed typo in match
The previous match was not matching `The tarantula is currently set to consume Alchemy experience.`
`/current set to consume .* experience\./`

Updated to:
`/current set to consume .* experience\./`

This fix matters for 1st use of a new tarantula.  I still need to go test it out on the test server on a character without one to make another change.