[script][craft.lic] - Quick fix for enchanting discipline
As reported by Karathos in Discord, the script is broken for the enchanting discipline.

It passes a variable `info` into `DRCC.check_for_existing_sigil?`, but I don't see info defined locally within this def. Other calls to that method pass a simple `stock_room` variable, which looks for the stock room of your hometown.

In the other crafting disciplines, info is a variable that is populated with craft_overrides from your yaml. That variable is never created or populated in the enchanting discipline section of craft. I don't know whether that's on purpose. I'm too weak at crafting to know. But this is at least a fix to something currently broken.

Karathos tested this fix.