add pass through
Since the new methods answer a question, they ought to reflect that. However, changing the method name before changing all instances of the method's use to the new name would be a problem. This method is used a LOT, both on and off the repo, so adding the pass through would make it backwards compatibible while giving the method a name that matches it's new attributes.