changed to generic give
Issue here was boxes with very similar or same names were being passed incorrectly eg iron skippet and ironwood skippet. You'd end up passing the freshly opened box back and hang on the open for the still-locked box. Dropping the item for the give_item gives what is in our right hand, which is always the next box, then processes the one in our left hand, which is always the one just opened.