enchant - Added messaging to stamp bput
When stamping an enchanted item (at least the one I'm making currently), it looks like this:
`You lazily wave the stamp over the freshly enchanted totem.  The imbue spell's expiring magic latches onto the engraved image and permanently imprint it within the sigil lines.`

I added a match to the bput for the first part of that.