[script][smoke] Add new images, mastery mid-loop
Adds the names of the new images from GF2023 (tentative) and sorts(hopefully) an issue where you've begun a loop and mastered everything mid-loop. Normally this would lead to just exhaling a line and a missed match. Instead, we'll just load the full list and sample for the duration of our loop.