Update task-forage.lic to handle cluttered rooms
Added new optional YAML setting called "wait_in_place." When true, to kill time the script will park in front of the task giver and collect rocks. When false, it runs to the outdoors_room and runs outdoorsmanship.lic.

If clutter is detected while waiting in place, the player is forced to wait_in_place = false.

If clutter is detected while gathering task items, the next closest room is attempted.