Add files via upload
DragonRealms Map Color Inversion Script

This is a Python script that runs through a directory and inverts all images
in the directory and then applies a color swap function to change one single
color into a different color.  This script is intended for changing DragonRealms
maps in Lich into a dark theme.