[script] sort-scrolls - some major updates
I've changed from using loops with `line = get` to `Lich::Util::issue_command()` as it is more efficient and reduces the complexity of the code.

I changed from using bputs to get the cases and scrolls out to `DRCI.get_item_safe?()`.

I added the abbreviation of spells to the big array we build at the start of the script and also added a check for when you can't turn to a full spell name.  Now it will try to turn to the abbreviation if the full name fails.

I updated from using `inv container` and trying to capture any line that has a scroll in it to now using `DRCI.get_scroll_list_in_container(@default_container)` which uses the `rummage /SC` option to find only scrolls in that container.

I think that's all.