gondola information
Added a simple `look gondola` command if the gondola is not there for informational purposes if paying attention to the script so you don't have to manually type it to know where the gondola is. 

No change to the movement functionality of the script. This is simply to facilitate laziness in the event the user wants information on the gondola's location.