[script] [combat-trainer] Added necro_corpse_priority: setting
Added `necro_corpse_priority:` setting to allow for prioritization of using corpses for either healing or pet creation.

Default behavior is to prioritze healing from corpses. Exclude it, leave it blank or set it to `heal`.
To have it prioritize pet creation from corpses set it to `pet`.

Rituals other than consue/arise have not changed. They are handled separately in the rituals section and continue to be prioritized behind pets and healing as before.